+++ named-entity-recognition/tests/compile_proto.py
