+++ named-entity-recognition/tests/path_setup.py
