--- sentiment-analysis/test/test_data/b64_sentences.py
@@ -1,2 +1,0 @@
-def senteces():
-    return "QmFzZWQgb24gYW4gYWN0dWFsIHN0b3J5LCBKb2huIEJvb3JtYW4gc2hvd3MgdGhlIHN0cnVnZ2xlIG9mIGFuIEFtZXJpY2FuIGRvY3Rvciwgd2hvc2UgaHVzYmFuZCBhbmQgc29uIHdlcmUgbXVyZGVyZWQgYW5kIHNoZSB3YXMgY29udGludWFsbHkgcGxhZ3VlZCB3aXRoIGhlciBsb3NzLiBBIGhvbGlkYXkgdG8gQnVybWEgd2l0aCBoZXIgc2lzdGVyIHNlZW1lZCBsaWtlIGEgZ29vZCBpZGVhIHRvIGdldCBhd2F5IGZyb20gaXQgYWxsLCBidXQgd2hlbiBoZXIgcGFzc3BvcnQgd2FzIHN0b2xlbiBpbiBSYW5nb29uLCBzaGUgY291bGQgbm90IGxlYXZlIHRoZSBjb3VudHJ5IHdpdGggaGVyIHNpc3RlciwgYW5kIHdhcyBmb3JjZWQgdG8gc3RheSBiYWNrIHVudGlsIHNoZSBjb3VsZCBnZXQgSS5ELiBwYXBlcnMgZnJvbSB0aGUgQW1lcmljYW4gZW1iYXNzeS4gVG8gZmlsbCBpbiBhIGRheSBiZWZvcmUgc2hlIGNvdWxkIGZseSBvdXQsIHNoZSB0b29rIGEgdHJpcCBpbnRvIHRoZSBjb3VudHJ5c2lkZSB3aXRoIGEgdG91ciBndWlkZS4gIkkgdHJpZWQgZmluZGluZyBzb21ldGhpbmcgaW4gdGhvc2Ugc3RvbmUgc3RhdHVlcywgYnV0IG5vdGhpbmcgc3RpcnJlZCBpbiBtZS4gSSB3YXMgc3RvbmUgbXlzZWxmLiIgPGJyIC8+PGJyIC8+U3VkZGVubHkgYWxsIGhlbGwgYnJva2UgbG9vc2UgYW5kIHNoZSB3YXMgY2F1Z2h0IGluIGEgcG9saXRpY2FsIHJldm9sdC4gSnVzdCB3aGVuIGl0IGxvb2tlZCBsaWtlIHNoZSBoYWQgZXNjYXBlZCBhbmQgc2FmZWx5IGJvYXJkZWQgYSB0cmFpbiwgc2hlIHNhdyBoZXIgdG91ciBndWlkZSBnZXQgYmVhdGVuIGFuZCBzaG90LiBJbiBhIHNwbGl0IHNlY29uZCBzaGUgZGVjaWRlZCB0byBqdW1wIGZyb20gdGhlIG1vdmluZyB0cmFpbiBhbmQgdHJ5IHRvIHJlc2N1ZSBoaW0sIHdpdGggbm8gdGhvdWdodCBvZiBoZXJzZWxmLiBDb250aW51YWxseSBoZXIgbGlmZSB3YXMgaW4gZGFuZ2VyLiA8YnIgLz48YnIgLz5IZXJlIGlzIGEgd29tYW4gd2hvIGRlbW9uc3RyYXRlZCBzcG9udGFuZW91cywgc2VsZmxlc3MgY2hhcml0eSwgcmlza2luZyBoZXIgbGlmZSB0byBzYXZlIGFub3RoZXIuIFBhdHJpY2lhIEFycXVldHRlIGlzIGJlYXV0aWZ1bCwgYW5kIG5vdCBqdXN0IHRvIGxvb2sgYXQ7IHNoZSBoYXMgYSBiZWF1dGlmdWwgaGVhcnQuIFRoaXMgaXMgYW4gdW5mb3JnZXR0YWJsZSBzdG9yeS4gPGJyIC8+PGJyIC8+IldlIGFyZSB0YXVnaHQgdGhhdCBzdWZmZXJpbmcgaXMgdGhlIG9uZSBwcm9taXNlIHRoYXQgbGlmZSBhbHdheXMga2VlcHMuIgoKVGhpcyBpcyBhIGdlbS4gQXMgYSBGaWxtIEZvdXIgcHJvZHVjdGlvbiAtIHRoZSBhbnRpY2lwYXRlZCBxdWFsaXR5IHdhcyBpbmRlZWQgZGVsaXZlcmVkLiBTaG90IHdpdGggZ3JlYXQgc3R5bGUgdGhhdCByZW1pbmRlZCBtZSBzb21lIEVycm9sIE1vcnJpcyBmaWxtcywgd2VsbCBhcnJhbmdlZCBhbmQgc2ltcGx5IGdyaXBwaW5nLiBJdCdzIGxvbmcgeWV0IGhvcnJpZnlpbmcgdG8gdGhlIHBvaW50IGl0J3MgZXhjcnVjaWF0aW5nLiBXZSBrbm93IHNvbWV0aGluZyBiYWQgaGFwcGVuZWQgKG9uZSBjYW4gZ3Vlc3MgYnkgdGhlIGxhY2sgb2YgcGFydGljaXBhdGlvbiBvZiBhIHBlcnNvbiBpbiB0aGUgaW50ZXJ2aWV3cykgYnV0IHdlIGFyZSBjb21wZWxsZWQgdG8gc2VlIGl0LCBhIGJpdCBsaWtlIGEgY2FyIGFjY2lkZW50IGluIHNsb3cgbW90aW9uLiBUaGUgc3Rvcnkgc3BhbnMgbW9zdCBjb25jZWl2YWJsZSBhc3BlY3RzIGFuZCB1bmxpa2Ugc29tZSBkb2N1bWVudGFyaWVzIGRpZCBub3QgdHJ5IGFuZCByZWZyYWluIGZyb20gc2hvd2luZyB0aGUgZ3JpbW1lciBzaWRlcyBvZiB0aGUgc3RvcmllcywgYXMgYWxzbyBkZWFsaW5nIHdpdGggdGhlIGd1aWx0IG9mIHRoZSBwZW9wbGUgRG9uIGxlZnQgYmVoaW5kIGhpbSwgd29uZGVyaW5nIHdoeSB0aGV5IGRpZG4ndCBzdG9wIGhpbSBpbiB0aW1lLiBJdCB0b29rIG1lIGEgZmV3IGhvdXJzIHRvIGdldCBvdXQgb2YgdGhlIG1lbGFuY2hvbHkgdGhhdCBncmlwcGVkIG1lIGFmdGVyIHNlZWluZyB0aGlzIHZlcnktd2VsbCBtYWRlIGRvY3VtZW50YXJ5LgoKSSByZWFsbHkgbGlrZSB0aGlzIHNob3cuIEl0IGhhcyBkcmFtYSwgcm9tYW5jZSwgYW5kIGNvbWVkeSBhbGwgcm9sbGVkIGludG8gb25lLiBJIGFtIDI4IGFuZCBJIGFtIGEgbWFycmllZCBtb3RoZXIsIHNvIEkgY2FuIGlkZW50aWZ5IGJvdGggd2l0aCBMb3JlbGVpJ3MgYW5kIFJvcnkncyBleHBlcmllbmNlcyBpbiB0aGUgc2hvdy4gSSBoYXZlIGJlZW4gd2F0Y2hpbmcgbW9zdGx5IHRoZSByZXBlYXRzIG9uIHRoZSBGYW1pbHkgQ2hhbm5lbCBsYXRlbHksIHNvIEkgYW0gbm90IHVwLXRvLWRhdGUgb24gd2hhdCBpcyBnb2luZyBvbiBub3cuIEkgdGhpbmsgZmVtYWxlcyB3b3VsZCBsaWtlIHRoaXMgc2hvdyBtb3JlIHRoYW4gbWFsZXMsIGJ1dCBJIGtub3cgc29tZSBtZW4gb3V0IHRoZXJlIHdvdWxkIGVuam95IGl0ISBJIHJlYWxseSBsaWtlIHRoYXQgaXMgYW4gaG91ciBsb25nIGFuZCBub3QgYSBoYWxmIGhvdXIsIGFzIHRoIGhvdXIgc2VlbXMgdG8gZmx5IGJ5IHdoZW4gSSBhbSB3YXRjaGluZyBpdCEgR2l2ZSBpdCBhIGNoYW5jZSBpZiB5b3UgaGF2ZSBuZXZlciBzZWVuIHRoZSBzaG93ISBJIHRoaW5rIExvcmVsZWkgYW5kIEx1a2UgYXJlIG15IGZhdm9yaXRlIGNoYXJhY3RlcnMgb24gdGhlIHNob3cgdGhvdWdoLCBtYWlubHkgYmVjYXVzZSBvZiB0aGUgd2F5IHRoZXkgYXJlIHdpdGggb25lIGFub3RoZXIuIEhvdyBjb3VsZCB5b3Ugbm90IHNlZSBzb21ldGhpbmcgd2FzIHRoZXJlIChvciB0YWtlIHRoYXQgbG9uZyB0byBzZWUgaXQgSSBndWVzcyBJIHNob3VsZCBzYXkpPyA8YnIgLz48YnIgLz5IYXBweSB2aWV3aW5nIQoKQWxhbiBSaWNrbWFuICYgRW1tYSBUaG9tcHNvbiBnaXZlIGdvb2QgcGVyZm9ybWFuY2VzIHdpdGggc291dGhlcm4vTmV3IE9ybGVhbnMgYWNjZW50cyBpbiB0aGlzIGRldGVjdGl2ZSBmbGljay4gSXQncyB3b3J0aCBzZWVpbmcgZm9yIHRoZWlyIHNjZW5lcy0gYW5kIFJpY2ttYW4ncyBzY2VuZSB3aXRoIEhhbCBIb2xicm9vay4gVGhlc2UgdGhyZWUgYWN0b3JzIG1hbm5hZ2UgdG8gZW50ZXJ0YWluIHVzIG5vIG1hdHRlciB3aGF0IHRoZSBtb3ZpZSwgaXQgc2VlbXMuIFRoZSBwbG90IGZvciB0aGUgbW92aWUgc2hvd3MgcG90ZW50aWFsLCBidXQgb25lIGdldHMgdGhlIGltcHJlc3Npb24gaW4gd2F0Y2hpbmcgdGhlIGZpbG0gdGhhdCBpdCB3YXMgbm90IHB1bGxlZCBvZmYgYXMgd2VsbCBhcyBpdCBjb3VsZCBoYXZlIGJlZW4uIFRoZSBmYWN0IHRoYXQgaXQgaXMgY2x1dHRlcmVkIGJ5IGEgcmF0aGVyIHVuaW50ZXJlc3Rpbmcgc3VicGxvdCBhbmQgbW9zdGx5IHVuaW50ZXJlc3Rpbmcga2lkbmFwcGVycyByZWFsbHkgbXVkZGxlcyB0aGluZ3MuIFRoZSBtb3ZpZSBpcyB3b3J0aCBhIHZpZXctIGlmIGZvciBub3RoaW5nIG1vcmUgdGhhbiBlbnRlcnRhaW5pbmcgcGVyZm9ybWFuY2VzIGJ5IFJpY2ttYW4sIFRob21wc29uLCBhbmQgSG9sYnJvb2suCgpJIGhhdmUgc2VlbiB0aGlzIG1vdmllIGFuZCBJIGRpZCBub3QgY2FyZSBmb3IgdGhpcyBtb3ZpZSBhbnlob3cuIEkgd291bGQgbm90IHRoaW5rIGFib3V0IGdvaW5nIHRvIFBhcmlzIGJlY2F1c2UgSSBkbyBub3QgbGlrZSB0aGlzIGNvdW50cnkgYW5kIGl0cyBuYXRpb25hbCBjYXBpdGFsLiBJIGRvIG5vdCBsaWtlIHRvIGxlYXJuIGZyZW5jaCBhbnlob3cgYmVjYXVzZSBJIGRvIG5vdCB1bmRlcnN0YW5kIHRoZWlyIGxhbmd1YWdlLiBXaHkgd291bGQgSSBnbyB0byBGcmFuY2Ugd2hlbiBJIHJhdGhlciBnbyB0byBHZXJtYW55IG9yIHRoZSBVbml0ZWQgS2luZ2RvbT8gR2VybWFueSBhbmQgdGhlIFVuaXRlZCBLaW5nZG9tIGFyZSB0aGUgbmF0aW9ucyBJIHRvbGVyYXRlLiBBcHBhcmVudGx5IHRoZSBPbHNlbiBUd2lucyBkbyBub3QgdW5kZXJzdGFuZCB0aGUgRnJlbmNoIGxhbmd1YWdlIGp1c3QgbGlrZSBtZS4gVGhlcmVmb3JlIEkgd2lsbCBub3QgYm90aGVyIHRoZSBGcmFuY2UgdHJpcCBubyBtYXR0ZXIgd2hhdC4gSSBtaWdodCBhcyB3ZWxsIHN0aWNrIHRvIHRoZSBVbml0ZWQgS2luZ2RvbSBhbmQgbWVldCBzaW5nbGUgd29tZW4gYW5kIHBsYXkgdmlkZW8gZ2FtZXMgaWYgdGhlcmUgaXMgYSB2aWRlbyBhcmNhZGUuIFRoYXQgaXMgYWxsLgoKSW4gTG9zIEFuZ2VsZXMsIHRoZSBhbGNvaG9saWMgYW5kIGxhenkgSGFuayBDaGluYXNraSAoTWF0dCBEaWxsb24pIHBlcmZvcm1zIGEgd2lkZSByYW5nZSBvZiBub24tcXVhbGlmaWVkIGZ1bmN0aW9ucyBqdXN0IHRvIGdldCBlbm91Z2ggbW9uZXkgdG8gZHJpbmsgYW5kIGdhbWJsZSBpbiBob3JzZSByYWNlcy4gSGlzIHByaW1hcnkgYW5kIG9ubHkgb2JqZWN0aXZlIGlzIHdyaXRpbmcgYW5kIGhhdmluZyBzZXh5IHdpdGggZGlydHkgd29tZW4uPGJyIC8+PGJyIC8+IkZhY3RvdHVtIiBpcyBhbiB1bmludGVyZXN0aW5nLCBwb2ludGxlc3MgYW5kIGV4dHJlbWVseSBib3JpbmcgbW92aWUgYWJvdXQgYW4gaXJyZXNwb25zaWJsZSBkcnVua2VuIHZhZ3JhbnQgdGhhdCB3b3JrcyBhIGNvdXBsZSBvZiBkYXlzIG9yIHdlZWtzIGp1c3QgdG8gZ2V0IGVub3VnaCBtb25leSB0byBidXkgc3Bpcml0cyBhbmQgZ2FtYmxlLCBiZWluZyBpbW1lZGlhdGVseSBmaXJlZCBkdWUgdG8gaGlzIHJlY2tsZXNzIGJlaGF2aW9yLiBJbiBhY2NvcmRhbmNlIHdpdGggSU1EYiwgdGhpcyBjaGFyYWN0ZXIgd291bGQgYmUgdGhlIGZpY3Rpb25hbCBhbHRlci1lZ28gb2YgdGhlIGF1dGhvciBDaGFybGVzIEJ1a293c2tpLCBhbmQgYmFzZWQgb24gdGhpcyBzdG9yeSwgSSB3aWxsIGNlcnRhaW5seSBuZXZlciByZWFkIGFueSBvZiBoaXMgbm92ZWxzLiBIb25lc3RseSwgaWYgdGhlIHZpZXdlciBsaWtlcyB0aGlzIHRoZW1lIG9mIGFsY29ob2xpYyBjb3VwbGVzLCBiZXR0ZXIgb2ZmIHdhdGNoaW5nIHRoZSB0b3VjaGluZyBhbmQgaGVhcnRicmVha2luZyBIZWN0b3IgQmFiZW5jbydzICJJcm9ud2VlZCIgb3IgTWFyY28gRmVycmVyaSdzICJTdG9yaWUgZGkgT3JkaW5hcmlhIEZvbGxpYSIgdGhhdCBpcyBiYXNlZCBvbiB0aGUgbGlmZSBvZiB0aGUgc2FtZSB3cml0ZXIuIE15IHZvdGUgaXMgZm91ci48YnIgLz48YnIgLz5UaXRsZSAoQnJhemlsKTogIkZhY3RvdHVtIMKWIFNlbSBEZXN0aW5vIiAoIkZhY3RvdHVtIMKWIFdpdGhvdXQgRGVzdGlueSIp"
