+++ sentiment-analysis/tests/compile_proto.py
