+++ sentiment-analysis/tests/path_setup.py
