+++ sentiment-analysis/tests/test_data/b64_sentences.py
@@ -1,0 +1,2 @@
+def senteces():
+    return "QnJvbXdlbGwgSGlnaCBpcyBhIGNhcnRvb24gY29tZWR5LiBJdCByYW4gYXQgdGhlIHNhbWUgdGltZSBhcyBzb21lIG90aGVyIHByb2dyYW1zIGFib3V0IHNjaG9vbCBsaWZlLCBzdWNoIGFzICJUZWFjaGVycyIuIE15IDM1IHllYXJzIGluIHRoZSB0ZWFjaGluZyBwcm9mZXNzaW9uIGxlYWQgbWUgdG8gYmVsaWV2ZSB0aGF0IEJyb213ZWxsIEhpZ2gncyBzYXRpcmUgaXMgbXVjaCBjbG9zZXIgdG8gcmVhbGl0eSB0aGFuIGlzICJUZWFjaGVycyIuIFRoZSBzY3JhbWJsZSB0byBzdXJ2aXZlIGZpbmFuY2lhbGx5LCB0aGUgaW5zaWdodGZ1bCBzdHVkZW50cyB3aG8gY2FuIHNlZSByaWdodCB0aHJvdWdoIHRoZWlyIHBhdGhldGljIHRlYWNoZXJzJyBwb21wLCB0aGUgcGV0dGluZXNzIG9mIHRoZSB3aG9sZSBzaXR1YXRpb24sIGFsbCByZW1pbmQgbWUgb2YgdGhlIHNjaG9vbHMgSSBrbmV3IGFuZCB0aGVpciBzdHVkZW50cy4gV2hlbiBJIHNhdyB0aGUgZXBpc29kZSBpbiB3aGljaCBhIHN0dWRlbnQgcmVwZWF0ZWRseSB0cmllZCB0byBidXJuIGRvd24gdGhlIHNjaG9vbCwgSSBpbW1lZGlhdGVseSByZWNhbGxlZCBhdCBIaWdoLiBBIGNsYXNzaWMgbGluZTogSU5TUEVDVE9SOiBJJ20gaGVyZSB0byBzYWNrIG9uZSBvZiB5b3VyIHRlYWNoZXJzLiBTVFVERU5UOiBXZWxjb21lIHRvIEJyb213ZWxsIEhpZ2guIEkgZXhwZWN0IHRoYXQgbWFueSBhZHVsdHMgb2YgbXkgYWdlIHRoaW5rIHRoYXQgQnJvbXdlbGwgSGlnaCBpcyBmYXIgZmV0Y2hlZC4gV2hhdCBhIHBpdHkgdGhhdCBpdCBpc24ndCEKClN0b3J5IG9mIGEgbWFuIHdobyBoYXMgdW5uYXR1cmFsIGZlZWxpbmdzIGZvciBhIHBpZy4gU3RhcnRzIG91dCB3aXRoIGEgb3BlbmluZyBzY2VuZSB0aGF0IGlzIGEgdGVycmlmaWMgZXhhbXBsZSBvZiBhYnN1cmQgY29tZWR5LiBBIGZvcm1hbCBvcmNoZXN0cmEgYXVkaWVuY2UgaXMgdHVybmVkIGludG8gYW4gaW5zYW5lLCB2aW9sZW50IG1vYiBieSB0aGUgY3JhenkgY2hhbnRpbmdzIG9mIGl0J3Mgc2luZ2Vycy4gVW5mb3J0dW5hdGVseSBpdCBzdGF5cyBhYnN1cmQgdGhlIFdIT0xFIHRpbWUgd2l0aCBubyBnZW5lcmFsIG5hcnJhdGl2ZSBldmVudHVhbGx5IG1ha2luZyBpdCBqdXN0IHRvbyBvZmYgcHV0dGluZy4gRXZlbiB0aG9zZSBmcm9tIHRoZSBlcmEgc2hvdWxkIGJlIHR1cm5lZCBvZmYuIFRoZSBjcnlwdGljIGRpYWxvZ3VlIHdvdWxkIG1ha2UgU2hha2VzcGVhcmUgc2VlbSBlYXN5IHRvIGEgdGhpcmQgZ3JhZGVyLiBPbiBhIHRlY2huaWNhbCBsZXZlbCBpdCdzIGJldHRlciB0aGFuIHlvdSBtaWdodCB0aGluayB3aXRoIHNvbWUgZ29vZCBjaW5lbWF0b2dyYXBoeSBieSBmdXR1cmUgZ3JlYXQgVmlsbW9zIFpzaWdtb25kLiBGdXR1cmUgc3RhcnMgU2FsbHkgS2lya2xhbmQgYW5kIEZyZWRlcmljIEZvcnJlc3QgY2FuIGJlIHNlZW4gYnJpZWZseS4="
